v2.3 retrieve children exit status

+++ src/main.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cecompte <cecompte@student.42.fr>          +#+  +:+       +#+        */
+/*   By: user <user@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/09 11:13:28 by cecompte          #+#    #+#             */
-/*   Updated: 2025/09/12 18:28:12 by cecompte         ###   ########.fr       */
+/*   Updated: 2025/09/22 18:31:07 by user             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,25 +19,39 @@ void	open_files(t_ids *id, char **argv)
 	{
 		perror(NULL);
 		id->tmp = 1;
-		id->fd[0] = open(argv[1], O_CREAT | O_RDONLY);
 	}
-	else
-		id->fd[0] = open(argv[1], O_RDONLY);
-	id->fd[1] = open(argv[4], O_CREAT | O_RDWR | O_TRUNC, 0644);
+	id->fd[0] = open(argv[1], O_CREAT | O_RDONLY);
 	if (id->fd[0] < 0)
 		exit_error(1);
+	id->fd[1] = open(argv[4], O_CREAT | O_RDWR | O_TRUNC, 0644);
 	if (id->fd[1] < 0)
 	{
 		close(id->fd[0]);
-		unlink(argv[1]);
+		if (id->tmp == 1)
+			unlink(argv[1]);
 		exit_error(1);
 	}
+}
+
+int	parent(t_ids id)
+{
+	int		status1;
+	int		status2;
+
+	waitpid(id.child1, &status1, 0);
+    waitpid(id.child2, &status2, 0);
+    if (WIFEXITED(status1) && WEXITSTATUS(status1) != 0) {
+        return (WEXITSTATUS(status1));
+    }
+    if (WIFEXITED(status2) && WEXITSTATUS(status2) != 0) {
+        return (WEXITSTATUS(status2));
+    }
+	return (0);
 }
 
 int	main(int argc, char **argv, char **envp)
 {
 	t_ids	id;
-	int		status;
 
 	if (argc != 5)
 		return (ft_putstr_fd("Wrong nb of arguments\n", 2), 1);
@@ -57,9 +71,8 @@ int	main(int argc, char **argv, char **envp)
 	}
 	close_all(id);
 	if (id.tmp == 1)
-		unlink(argv[1]);
-	waitpid(id.child1, &status, 0);
-	waitpid(id.child2, &status, 0);
+		return (unlink(argv[1]), 1);
+	return (parent(id));
 }
 
 /* to do
